RPRBLND-1813: World support (#80)
PURPOSE

World export implementation is outdated and was working in draft mode. We need more complex support of world nodes, make it work in USD nodes and in viewport.

EFFECT OF CHANGE

Improved worlds export system with better world nodes support, make it work in USD nodes and available to change in viewport.